CRUD y Controller de EstadoCivil, Aseguradora, Género, Pais y TipoIdentificador: Completo y andando.

+++ Controllers/AseguradoraController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using System;
 using System.Threading.Tasks;
 using UTNApiTalleres.Data.Repositorio.Interfaz;
 using WebApiTalleres.Models;
@@ -20,13 +21,36 @@ namespace UTNApiTalleres.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAllAseguradora()
         {
-            return Ok(await _aseguradoraDao.findAll());
+            try
+            {                
+                var aseguradoras = await _aseguradoraDao.findAll();
+               
+                return Ok(aseguradoras);
+            
+            }
+            catch (Exception ex)
+            {
+                //log error
+                return StatusCode(500, ex.Message);
+            }
+          
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetAseguradora(int id)
         {
-            return Ok(await _aseguradoraDao.find(id));
+            try
+            {
+                var aseguradora =  await _aseguradoraDao.find(id);
+                if (aseguradora == null)
+                    return NotFound();
+                return Ok(aseguradora);
+            }
+            catch (Exception ex)
+            {
+                //log error
+                return StatusCode(500, ex.Message);
+            }
         }
 
         [HttpPost]
@@ -40,26 +64,60 @@ namespace UTNApiTalleres.Controllers
 
             var oCreated = await _aseguradoraDao.create(aseguradora);
             
-            return Created("created", oCreated);
+            return Created("Se creo exitosamente.", oCreated);
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateAseguradora(int id)
-        {
-        
-             await _aseguradoraDao.delete(id);
+        public async Task<IActionResult> UpdateAseguradora([FromBody] Aseguradora Aseguradora)
+        {                     
 
-            return NoContent();
+            try
+            {
+                //validamos que exista la persona
+                var aseguradora = await _aseguradoraDao.find(Aseguradora.Id);
+
+                //en caso de no existir retornamos Not Found
+                if (aseguradora == null)
+                    return NotFound("La Aseguradora no existe.");
+
+                //En caso de existir avanzamos con la actualización
+                var actualizado = await _aseguradoraDao.update(Aseguradora);
+
+                if (actualizado)
+                    return Content("Actualización exitosa.");
+                else
+                    return NoContent();
+            }
+            catch (Exception ex)
+            {
+                //log error
+                return StatusCode(500, ex.Message);
+            }
         }
 
-        [HttpDelete]
+        [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAseguradora(int id)
         {
-         
 
-            await _aseguradoraDao.delete(id);
+            try
+            {
 
-            return NoContent();
+                var aseguradora = await _aseguradoraDao.find(id);
+
+                if (aseguradora == null)
+                    return NotFound("Aseguradora no encontrada.");
+
+                await _aseguradoraDao.delete(id);
+
+                return NoContent();
+
+            }
+            catch (Exception ex)
+            {
+                //log error
+                return StatusCode(500, ex.Message);
+            }
+     
         }
 
     }
